refactor: dialog edit method
