+++ lib/app/view_model/post_state.dart
@@ -5,7 +5,9 @@ import 'package:mvvm_pattern/app/model/infra/firebase_provider.dart';
 import 'package:mvvm_pattern/app/model/post/post.dart';
 
 // generaterを使ったほうがもっと楽に書ける
-final postStateAsyncProvider =  AsyncNotifierProvider<PostAsyncNotifier, void>(PostAsyncNotifier.new);
+final postStateAsyncProvider =
+    AsyncNotifierProvider<PostAsyncNotifier, void>(PostAsyncNotifier.new);
+
 // 非同期のデータを扱うので、AutoDisposeAsyncNotifierを使う
 class PostAsyncNotifier extends AsyncNotifier<void> {
   @override
@@ -15,8 +17,9 @@ class PostAsyncNotifier extends AsyncNotifier<void> {
 
   Future<void> addPost(Post post) async {
     final postRef = ref.read(postReferenceWithConverter);
-    state = const AsyncLoading();// 通信中の状態をセット
-    state = await AsyncValue.guard(() async {// 通信結果をセット
+    state = const AsyncLoading(); // 通信中の状態をセット
+    state = await AsyncValue.guard(() async {
+      // 通信結果をセット
       await postRef.add(post);
     });
   }
